Stitch chunks into full valid events
It's unfortunately possible for events to spread across multiple HTTP
chunks in a chunked-encoded response. We have to be careful when reading
the streamed chunked response as the end of the chunk might not be the
end of an event.

We have to look for the event delimiter (empty new line) and stitch back
together consecutive chunks into individual events as/when necessary.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>